feat: passing analytics and database to Addition and Removal through interfaces
